# docs/swagger.yaml
info:
  contact: {}
paths:
  /admin/category-create:
    post:
      description: 创建分类
      parameters:
      - description: authorization
        in: header
        name: authorization
        required: true
        type: string
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: problem_ids
        in: formData
        name: problem_ids
        type: integer
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 创建分类
      tags:
      - 管理员私有方法
  /admin/category-delete:
    delete:
      description: 删除分类
      parameters:
      - description: authorization
        in: header
        name: authorization
        required: true
        type: string
      - description: identity
        in: query
        name: identity
        required: true
        type: string
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 删除分类
      tags:
      - 管理员私有方法
  /admin/category-list:
    get:
      description: 获取分类列表
      parameters:
      - description: authorization
        in: header
        name: authorization
        required: true
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: keyword
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取分类列表
      tags:
      - 管理员私有方法
  /admin/category-update:
    put:
      description: 更新分类
      parameters:
      - description: authorization
        in: header
        name: authorization
        required: true
        type: string
      - description: identity
        in: formData
        name: identity
        required: true
        type: string
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: ParentId
        in: formData
        name: ParentId
        type: integer
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 更新分类
      tags:
      - 管理员私有方法
  /admin/problem-create:
    post:
      consumes:
      - multipart/form-data
      description: 创建问题
      parameters:
      - description: authorization
        in: header
        name: authorization
        required: true
        type: string
      - description: title
        in: formData
        name: title
        required: true
        type: string
      - description: content
        in: formData
        name: content
        required: true
        type: string
      - description: max_mem
        in: formData
        name: max_mem
        type: integer
      - description: max_runtime
        in: formData
        name: max_runtime
        type: integer
      - collectionFormat: csv
        description: category_ids
        in: formData
        items:
          type: string
        name: category_ids
        type: array
      - collectionFormat: csv
        description: test_cases
        in: formData
        items:
          type: string
        name: test_cases
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 创建问题
      tags:
      - 管理员私有方法
  /admin/problem-update:
    put:
      description: 更新问题
      parameters:
      - description: authorization
        in: header
        name: authorization
        required: true
        type: string
      - description: identity
        in: formData
        name: identity
        required: true
        type: string
      - description: title
        in: formData
        name: title
        required: true
        type: string
      - description: content
        in: formData
        name: content
        required: true
        type: string
      - description: max_mem
        in: formData
        name: max_mem
        type: integer
      - description: max_runtime
        in: formData
        name: max_runtime
        type: integer
      - collectionFormat: csv
        description: category_ids
        in: formData
        items:
          type: string
        name: category_ids
        type: array
      - collectionFormat: csv
        description: test_cases
        in: formData
        items:
          type: string
        name: test_cases
        required: true
        type: array
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 更新问题
      tags:
      - 管理员私有方法
  /problem-detail:
    get:
      consumes:
      - application/json
      description: 获取问题详情
      parameters:
      - description: problem identity
        in: query
        name: identity
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取问题详情
      tags:
      - 公共方法
  /problem-list:
    get:
      consumes:
      - application/json
      description: 获取问题列表
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: category_identity
        in: query
        name: category_identity
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取问题列表
      tags:
      - 公共方法
  /user/detail:
    get:
      consumes:
      - application/json
      description: 获取用户详情
      parameters:
      - description: user identity
        in: query
        name: identity
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取用户详情
      tags:
      - 公共方法
  /user/login:
    post:
      description: 登录
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 登录
      tags:
      - 公共方法
  /user/rank-list:
    get:
      description: 获取排名列表
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取排名列表
      tags:
      - 公共方法
  /user/register:
    post:
      description: 用户注册
      parameters:
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      - description: 验证码
        in: formData
        name: code
        required: true
        type: string
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 手机号
        in: formData
        name: phone
        type: string
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - 公共方法
  /user/send-code:
    post:
      description: 发送验证码
      parameters:
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 发送验证码
      tags:
      - 公共方法
  /user/submit-list:
    get:
      consumes:
      - application/json
      description: 获取提交记录
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: problem_identity
        in: query
        name: problem_identity
        type: string
      - description: user_identity
        in: query
        name: user_identity
        type: string
      - description: status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\":200,\"data\":{\"count\":0,\"data\":[]}\'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取提交记录
      tags:
      - 公共方法
swagger: "2.0"